move package and refactor functions away from Store
